+++ Sources/MapboxNavigation/Route.swift
@@ -44,13 +44,13 @@ extension Route {
     }
     
     func congestionFeatures(legIndex: Int? = nil,
-                            isAlternativeRoute: Bool = false,
                             roadClassesWithOverriddenCongestionLevels: Set<MapboxStreetsRoadClass>? = nil) -> [Feature] {
         guard let coordinates = shape?.coordinates, let shape = shape else { return [] }
         var features: [Feature] = []
         
         for (index, leg) in legs.enumerated() {
             let legFeatures: [Feature]
+            let currentLegAttribute = (legIndex != nil) ? index == legIndex : true
             
             if let congestionLevels = leg.segmentCongestionLevels, congestionLevels.count < coordinates.count {
                 // The last coordinate of the preceding step, is shared with the first coordinate of the next step, we don't need both.
@@ -70,8 +70,7 @@ extension Route {
                     var feature = Feature(geometry: .lineString(LineString(congestionSegment.0)))
                     feature.properties = [
                         CongestionAttribute: String(describing: congestionSegment.1),
-                        "isAlternativeRoute": isAlternativeRoute,
-                        CurrentLegAttribute: (legIndex != nil) ? index == legIndex : true
+                        CurrentLegAttribute: currentLegAttribute
                     ]
                     
                     return feature
@@ -79,8 +78,7 @@ extension Route {
             } else {
                 var feature = Feature(geometry: .lineString(LineString(shape.coordinates)))
                 feature.properties = [
-                    "isAlternativeRoute": isAlternativeRoute,
-                    CurrentLegAttribute: (legIndex != nil) ? index == legIndex : true
+                    CurrentLegAttribute: currentLegAttribute
                 ]
                 legFeatures = [feature]
             }