+++ Tests/MapboxNavigationTests/NavigationMapViewTests.swift
@@ -127,7 +127,7 @@ class NavigationMapViewTests: TestCase {
 
         navigationMapView.initPrimaryRoutePoints(route: route)
         navigationMapView.updateUpcomingRoutePointIndex(routeProgress: testRouteProgress)
-        navigationMapView.updateTraveledRouteLine(targetPoint)
+        navigationMapView.updateFractionTraveled(coordinate: targetPoint)
 
         let expectedTraveledFraction = 0.06383308537010246
 
@@ -349,6 +349,60 @@ class NavigationMapViewTests: TestCase {
         congestions.enumerated().forEach {
             XCTAssertEqual(congestionLevel($0.element), expectedCongestionLevels[$0.offset])
         }
+    }
+    
+    func testUpdateRouteLineGradient() {
+        let route = loadRoute(from: "route-with-road-classes-single-congestion")
+        let congestions = route.congestionFeatures()
+        var lineGradient = navigationMapView.routeLineGradient(congestions, fractionTraveled: 0.0, isMain: true)
+        XCTAssertEqual(lineGradient[0.0], navigationMapView.trafficUnknownColor)
+        
+        var fractionTraveled = 0.5
+        var nextDownFractionTraveled = Double(CGFloat(fractionTraveled).nextDown)
+        lineGradient = navigationMapView.updateRouteLineGradientStops(fractionTraveled: fractionTraveled, gradientStops: lineGradient)
+
+        XCTAssertEqual(lineGradient[0.0], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[nextDownFractionTraveled], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[fractionTraveled], navigationMapView.trafficUnknownColor)
+        
+        let nexDownModerateFraction = Double(CGFloat(0.3).nextDown)
+        let nexDownHeavyFraction = Double(CGFloat(0.4).nextDown)
+        lineGradient = [
+            0.0: navigationMapView.trafficSevereColor,
+            nexDownModerateFraction: navigationMapView.trafficSevereColor,
+            0.3: navigationMapView.trafficModerateColor,
+            nexDownHeavyFraction: navigationMapView.trafficModerateColor,
+            0.4: navigationMapView.trafficHeavyColor
+        ]
+        
+        fractionTraveled = 0.3
+        nextDownFractionTraveled = Double(CGFloat(fractionTraveled).nextDown)
+        lineGradient = navigationMapView.updateRouteLineGradientStops(fractionTraveled: fractionTraveled, gradientStops: lineGradient)
+        XCTAssertEqual(lineGradient[0.0], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[nextDownFractionTraveled], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[fractionTraveled], navigationMapView.trafficModerateColor)
+        
+        fractionTraveled = 0.35
+        nextDownFractionTraveled = Double(CGFloat(fractionTraveled).nextDown)
+        lineGradient = navigationMapView.updateRouteLineGradientStops(fractionTraveled: fractionTraveled, gradientStops: lineGradient)
+        XCTAssertEqual(lineGradient[0.0], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[nextDownFractionTraveled], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[fractionTraveled], navigationMapView.trafficModerateColor)
+        
+        fractionTraveled = 0.45
+        nextDownFractionTraveled = Double(CGFloat(fractionTraveled).nextDown)
+        lineGradient = navigationMapView.updateRouteLineGradientStops(fractionTraveled: fractionTraveled, gradientStops: lineGradient)
+        XCTAssertEqual(lineGradient[0.0], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[nextDownFractionTraveled], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[fractionTraveled], navigationMapView.trafficHeavyColor)
+        
+        lineGradient = [Double:UIColor]()
+        fractionTraveled = 0.45
+        nextDownFractionTraveled = Double(CGFloat(fractionTraveled).nextDown)
+        lineGradient = navigationMapView.updateRouteLineGradientStops(fractionTraveled: fractionTraveled, gradientStops: lineGradient)
+        XCTAssertEqual(lineGradient[0.0], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[nextDownFractionTraveled], navigationMapView.traversedRouteColor)
+        XCTAssertEqual(lineGradient[fractionTraveled], navigationMapView.trafficUnknownColor)
     }
     
     func testRoadClassesWithOverriddenCongestionLevelsRemovesDuplicates() {